c program that append a text to a file

+++ 0x15-file_io/2-append_text_to_file.c
@@ -20,7 +20,7 @@ int append_text_to_file(const char *filename, char *text_content)
 
 	if (fd == -1)
 	{
-		return (-1);
+	return (-1);
 	}
 
 	if (text_content != NULL)
@@ -31,7 +31,7 @@ int append_text_to_file(const char *filename, char *text_content)
 
 		written = write(fd, text_content, lenght);
 
-		if (written == -1)
+		if (written == (size_t)-1)
 		{
 			close(fd);
 			return (-1);
